lot's of replacing `long` with `int` for many files

+++ MyHebrewBible/MyHebrewBible.Client/Components/VerseParagraphRange.cs
@@ -1,4 +1,4 @@
 ﻿namespace MyHebrewBible.Client.Components;
 
-public record VerseParagraphRange(long BegId, long EndId, string BCV);
+public record VerseParagraphRange(int BegId, int EndId, string BCV);
