+++ MyHebrewBible/MyHebrewBible.Client/Features/AlephTav/ScriptureIdAndVerseNumber.cs
@@ -1,3 +1,3 @@
 ﻿namespace MyHebrewBible.Client.Features.AlephTav;
 
-public record ScriptureIdAndVerseNumber(long ScriptureID, long VerseNumber);
+public record ScriptureIdAndVerseNumber(int ScriptureID, int VerseNumber);
