+++ MyHebrewBible/MyHebrewBible/Endpoints/GetAlephTavHebrewVerses.cs
@@ -4,8 +4,8 @@ namespace MyHebrewBible.Endpoints;
 
 public class WordPartByBookAndChapterRequest
 {
-	public long BookID { get; set; }
-	public long Chapter { get; set; }
+	public int BookID { get; set; }
+	public int Chapter { get; set; }
 }
 
 public class GetAlephTavHebrewVerses(Query query, ILogger<GetAlephTavHebrewVerses> logger) : Endpoint<WordPartByBookAndChapterRequest, IEnumerable<AlephTavHebrewVerse>>
@@ -41,20 +41,20 @@ public class GetAlephTavHebrewVerses(Query query, ILogger<GetAlephTavHebrewVerse
 
 public class AlephTavHebrewVerse
 {
-	public long ScriptureID { get; set; }
+	public int ScriptureID { get; set; }
 	public string? BCV { get; set; }
-	public long Chapter { get; set; }
-	public long Verse { get; set; }
-	public long WordCount { get; set; }
-	public long WordEnum { get; set; }
+	public int Chapter { get; set; }
+	public int Verse { get; set; }
+	public int WordCount { get; set; }
+	public int WordEnum { get; set; }
 	public string? Hebrew1 { get; set; }
 	public string? Hebrew2 { get; set; }
 	public string? Hebrew3 { get; set; }
 	public string? KjvWord { get; set; }
-	public long Strongs { get; set; }
+	public int Strongs { get; set; }
 	public string? Transliteration { get; set; }
 	public int? FinalEnum { get; set; }
-	public long HasTwo { get; set; }
+	public int HasTwo { get; set; }
 }
 
 // Ignore Spelling: BCV, strongs, Paseq, css,  alephtavhebrewverse: bookid,
